wasd

+++ main.py
@@ -33,51 +33,45 @@ Idling
 """ 
 
 def main():
-    # if True:
-    from spot_controller import SpotController
-    with SpotController(username=SPOT_USERNAME, password=SPOT_PASSWORD, robot_ip=ROBOT_IP) as spot:
+    if True:
+    # from spot_controller import SpotController
+    # with SpotController(username=SPOT_USERNAME, password=SPOT_PASSWORD, robot_ip=ROBOT_IP) as spot:
+        ws = create_connection("wss://6093-171-66-12-11.ngrok-free.app")
+        ws.send("Hello, World")
+        cam = CameraVideo(0, max_fps=10, height=360, width=480)
         while True:
             try:
-                ws = create_connection("wss://6093-171-66-12-11.ngrok-free.app")
-                ws.send("Hello, World")
-                cam = CameraVideo(0, max_fps=10, height=360, width=480)
-                while True:
+                cmd =  ws.recv()
+                if cmd == "[CAM]":
+                    frame = cam.get_frame()
+                    _, enc = cv2.imencode('.jpg', frame)
+                    jpg_as_text = base64.b64encode(enc)
+                    # print(jpg_as_text)
+                    ws.send(jpg_as_text)
+                    continue
+                if cmd == "[PAYLOAD]":
+                    with open("/tmp/payload.py", "w+") as f:
+                        f.seek(0)
+                        f.truncate()
+                        f.write(ws.recv())
                     try:
-                        cmd =  ws.recv()
-                        if cmd == "[CAM]":
-                            frame = cam.get_frame()
-                            _, enc = cv2.imencode('.jpg', frame)
-                            jpg_as_text = base64.b64encode(enc)
-                            # print(jpg_as_text)
-                            ws.send(jpg_as_text)
-                            continue
-                        if cmd == "[PAYLOAD]":
-                            with open("/tmp/payload.py", "w+") as f:
-                                f.seek(0)
-                                f.truncate()
-                                f.write(ws.recv())
-                            try:
-                                process = subprocess.Popen(["python3", "/tmp/payload.py"], stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
-                                for line in process.stdout:
-                                    ws.send(line.strip())
-                                for line in process.stderr:
-                                    ws.send(line.strip())
-                                process.wait()
-                            except Exception as e:
-                                ws.send(f"Error executing command: {e}")
-                            ws.send("[EOL]")
-                            continue
-                        res = eval(cmd)
-                        if res:
-                            ws.send(str(res))
-                        else:
-                            ws.send("None") 
+                        process = subprocess.Popen(["python3", "/tmp/payload.py"], stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
+                        for line in process.stdout:
+                            ws.send(line.strip())
+                        for line in process.stderr:
+                            ws.send(line.strip())
+                        process.wait()
                     except Exception as e:
-                        ws.send(str(e))
+                        ws.send(f"Error executing command: {e}")
+                    ws.send("[EOL]")
+                    continue
+                res = eval(cmd)
+                if res:
+                    ws.send(str(res))
+                else:
+                    ws.send("None") 
             except Exception as e:
-                print(e)
-                time.sleep(5)
-                continue
+                ws.send(str(e))
 
 
 if __name__ == '__main__':
